@@ -2,40 +2,42 @@
     "name": "laravel/laravel",
     "type": "project",
     "description": "The Laravel Framework.",
-    "keywords": [
-        "framework",
-        "laravel"
-    ],
+    "keywords": ["framework", "laravel"],
     "license": "MIT",
     "require": {
-        "php": "^8.0",
-        "astrotomic/laravel-translatable": "^11.9",
-        "bepsvpt/secure-headers": "^7.1",
-        "doctrine/dbal": "^3.2",
-        "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.0.1",
+        "php": "^8.1",
+        "archtechx/enums": "^0.3.1",
+        "bepsvpt/secure-headers": "^7.2",
+        "doctrine/dbal": "^3.4",
+        "guzzlehttp/guzzle": "^7.2",
+        "illuminate/console": "^9.29",
         "itsgoingd/clockwork": "^5.1",
         "kreait/laravel-firebase": "^3.3",
-        "laravel/framework": "^8.65",
-        "laravel/passport": "^10.2",
+        "laravel/framework": "^9.19",
+        "laravel/sanctum": "^3.0",
         "laravel/slack-notification-channel": "^2.3",
         "laravel/tinker": "^2.5",
-        "laravel/ui": "^3.4",
-        "league/flysystem-aws-s3-v3": "1.0",
+        "league/flysystem-aws-s3-v3": "^3.3",
         "mattkingshott/axiom": "^3.2",
-        "psr/cache": "2.0",
-        "psr/log": "1.*",
-        "rollbar/rollbar-laravel": "^7.0",
+        "psr/log": "~2",
+        "rollbar/rollbar-laravel": "^7.2",
         "spatie/eloquent-sortable": "^4.0",
-        "staudenmeir/belongs-to-through": "^2.11"
+        "spatie/laravel-translatable": "^6.0",
+        "spatie/laravel-translation-loader": "^2.7",
+        "staudenmeir/belongs-to-through": "^2.12"
     },
     "require-dev": {
-        "facade/ignition": "^2.5",
         "fakerphp/faker": "^1.9.1",
+        "laravel-lang/attributes": "^2.0",
+        "laravel-lang/lang": "^12.0",
+        "laravel-lang/publisher": "^14.3",
+        "laravel/pint": "^1.0",
         "laravel/sail": "^1.0.1",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^5.10",
-        "phpunit/phpunit": "^9.5.10"
+        "nunomaduro/collision": "^6.1",
+        "nunomaduro/larastan": "^2.0",
+        "phpunit/phpunit": "^9.5.10",
+        "spatie/laravel-ignition": "^1.0"
     },
     "autoload": {
         "psr-4": {
@@ -55,7 +57,7 @@
             "@php artisan package:discover --ansi"
         ],
         "post-update-cmd": [
-            "@php artisan vendor:publish --tag=laravel-assets --ansi"
+            "@php artisan vendor:publish --tag=laravel-assets --ansi --force"
         ],
         "post-root-package-install": [
             "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
@@ -72,7 +74,10 @@
     "config": {
         "optimize-autoloader": true,
         "preferred-install": "dist",
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "pestphp/pest-plugin": true
+        }
     },
     "minimum-stability": "dev",
     "prefer-stable": true
