@@ -2,26 +2,28 @@
 
 namespace Database\Factories;
 
+use App\Enums\Language;
 use Illuminate\Database\Eloquent\Factories\Factory;
-use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
 
+/**
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\User>
+ */
 class UserFactory extends Factory
 {
     /**
      * Define the model's default state.
      *
-     * @return array
+     * @return array<string, mixed>
      */
     public function definition()
     {
         return [
-            'first_name'        => $this->faker->name,
-            'last_name'         => $this->faker->name,
-            'gender'            => $this->faker->randomElement([1, 2, 9]),
-            'email'             => $this->faker->unique()->safeEmail,
+            'name'              => fake()->name(),
+            'email'             => fake()->unique()->safeEmail(),
+            'language'          => Language::English->value,
             'email_verified_at' => now(),
-            'password'          => Hash::make($this->faker->password()),
+            'password'          => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
             'remember_token'    => Str::random(10),
         ];
     }
@@ -29,14 +31,14 @@ class UserFactory extends Factory
     /**
      * Indicate that the model's email address should be unverified.
      *
-     * @return \Illuminate\Database\Eloquent\Factories\Factory
+     * @return static
      */
     public function unverified()
     {
-        return $this->state(function (array $attributes) {
-            return [
-                'email_verified_at' => null,
-            ];
-        });
+        return $this->state(fn (array $attributes) => [
+            'email_verified_at'                  => null,
+            'email_verification_code'            => rand(100000, 999999),
+            'email_verification_code_expires_at' => now()->addHour(),
+        ]);
     }
 }
