@@ -11,14 +11,18 @@ class Kernel extends ConsoleKernel
      * Define the application's command schedule.
      *
      * @param \Illuminate\Console\Scheduling\Schedule $schedule
+     *
+     * @return void
      */
     protected function schedule(Schedule $schedule)
     {
-        // $schedule->command('inspire')->hourly();
+        $schedule->command('sanctum:prune-expired --hours=24')->daily();
     }
 
     /**
      * Register the commands for the application.
+     *
+     * @return void
      */
     protected function commands()
     {
