@@ -2,43 +2,22 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
+use App\Http\Controllers\AuthenticatedController;
+use App\Http\Requests\Api\Maintenance\PostDownRequest;
+use App\Http\Requests\Api\Maintenance\PostUpRequest;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Artisan;
 
-class MaintenanceController extends Controller
+class MaintenanceController extends AuthenticatedController
 {
-    //======================================================================
-    // CONSTRUCTOR
-    //
-    // Current authenticated user is loaded into $this->user by the parent
-    // controller class
-    //
-    //======================================================================
-    public function __construct(Request $request)
-    {
-        parent::__construct();
-
-        $this->middleware(function ($request, $next) {
-            $this->authorize('maintenance', $this->user);
-
-            return $next($request);
-        });
-    }
-
-    //======================================================================
-    // ROUTER METHODS
-    //======================================================================
-
     /**
-     * Puts the api under maintenance mode.
+     * Puts the application down for maintenance.
      *
-     * @param Request $request
+     * @param PostDownRequest $request
      *
-     * @return array
+     * @return array<void>
      */
-    public function postEnable(Request $request)
+    public function postDown(PostDownRequest $request): array
     {
         if (!App::isDownForMaintenance()) {
             Artisan::call('down');
@@ -48,13 +27,13 @@ class MaintenanceController extends Controller
     }
 
     /**
-     * Removes the api from maintenance mode.
+     * Puts the application up from maintenance.
      *
-     * @param Request $request
+     * @param PostUpRequest $request
      *
-     * @return array
+     * @return array<void>
      */
-    public function postDisable(Request $request)
+    public function postUp(PostUpRequest $request): array
     {
         if (App::isDownForMaintenance()) {
             Artisan::call('up');
