@@ -11,12 +11,12 @@ class Kernel extends HttpKernel
      *
      * These middleware are run during every request to your application.
      *
-     * @var array
+     * @var array<int, class-string|string>
      */
     protected $middleware = [
         // \App\Http\Middleware\TrustHosts::class,
         \App\Http\Middleware\TrustProxies::class,
-        \Fruitcake\Cors\HandleCors::class,
+        \Illuminate\Http\Middleware\HandleCors::class,
         \App\Http\Middleware\PreventRequestsDuringMaintenance::class,
         \Illuminate\Foundation\Http\Middleware\ValidatePostSize::class,
         \App\Http\Middleware\TrimStrings::class,
@@ -27,27 +27,24 @@ class Kernel extends HttpKernel
     /**
      * The application's route middleware groups.
      *
-     * @var array
+     * @var array<string, array<int, class-string|string>>
      */
     protected $middlewareGroups = [
         'web' => [
             \App\Http\Middleware\EncryptCookies::class,
             \Illuminate\Cookie\Middleware\AddQueuedCookiesToResponse::class,
             \Illuminate\Session\Middleware\StartSession::class,
-            // \Illuminate\Session\Middleware\AuthenticateSession::class,
             \Illuminate\View\Middleware\ShareErrorsFromSession::class,
             \App\Http\Middleware\VerifyCsrfToken::class,
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
-            'permissions',
             'localize',
         ],
 
         'api' => [
+            // \Laravel\Sanctum\Http\Middleware\EnsureFrontendRequestsAreStateful::class,
             'throttle:api',
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
-            'permissions',
             'localize',
-            'lowercase_email',
         ],
     ];
 
@@ -56,26 +53,22 @@ class Kernel extends HttpKernel
      *
      * These middleware may be assigned to groups or used individually.
      *
-     * @var array
+     * @var array<string, class-string|string>
      */
     protected $routeMiddleware = [
         'auth'                => \App\Http\Middleware\Authenticate::class,
         'auth.basic'          => \Illuminate\Auth\Middleware\AuthenticateWithBasicAuth::class,
+        'auth.session'        => \Illuminate\Session\Middleware\AuthenticateSession::class,
         'cache.headers'       => \Illuminate\Http\Middleware\SetCacheHeaders::class,
         'can'                 => \Illuminate\Auth\Middleware\Authorize::class,
         'guest'               => \App\Http\Middleware\RedirectIfAuthenticated::class,
         'password.confirm'    => \Illuminate\Auth\Middleware\RequirePassword::class,
-        'signed'              => \Illuminate\Routing\Middleware\ValidateSignature::class,
+        'signed'              => \App\Http\Middleware\ValidateSignature::class,
         'throttle'            => \Illuminate\Routing\Middleware\ThrottleRequests::class,
         'verified'            => \Illuminate\Auth\Middleware\EnsureEmailIsVerified::class,
-        'permissions'         => \App\Http\Middleware\LoadPermissions::class,
         'localize'            => \App\Http\Middleware\Localize::class,
         'email_verified'      => \App\Http\Middleware\EmailVerified::class,
-        'sanitize_login'      => \App\Http\Middleware\SanitizeLogin::class,
-        'lowercase_email'     => \App\Http\Middleware\LowercaseEmail::class,
         'throttle_login'      => \App\Http\Middleware\ThrottleLogin::class,
-        'block_in_production' => \App\Http\Middleware\BlockInProduction::class,
-        'set_oauth_client'    => \App\Http\Middleware\SetOAuthClient::class,
-        'sanitize_refresh'    => \App\Http\Middleware\SanitizeRefresh::class,
+        'block-in-production' => \App\Http\Middleware\BlockInProduction::class,
     ];
 }
