@@ -9,10 +9,11 @@ class PreventRequestsDuringMaintenance extends Middleware
     /**
      * The URIs that should be reachable while maintenance mode is enabled.
      *
-     * @var array
+     * @var array<int, string>
      */
     protected $except = [
-        'maintenance/enable',
-        'maintenance/disable',
+        'auth/login',
+        'maintenance/up',
+        'maintenance/down',
     ];
 }
