@@ -2,23 +2,36 @@
 
 namespace App\Providers;
 
+use App\Models\PersonalAccessToken;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\ServiceProvider;
-use URL;
+use Illuminate\Validation\Rules\Password;
+use Laravel\Sanctum\Sanctum;
 
 class AppServiceProvider extends ServiceProvider
 {
     /**
      * Register any application services.
+     *
+     * @return void
      */
     public function register()
     {
+        Sanctum::ignoreMigrations();
+        Sanctum::usePersonalAccessTokenModel(PersonalAccessToken::class);
     }
 
     /**
      * Bootstrap any application services.
+     *
+     * @return void
      */
     public function boot()
     {
-        URL::forceScheme(config('app.scheme'));
+        Model::preventLazyLoading();
+
+        Password::defaults(function () {
+            return Password::min(8)->letters();
+        });
     }
 }
