@@ -2,26 +2,29 @@
 
 namespace App\Providers;
 
-use Carbon\Carbon;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
-use Illuminate\Support\Facades\Gate;
-use Laravel\Passport\Passport;
 
 class AuthServiceProvider extends ServiceProvider
 {
     /**
+     * The model to policy mappings for the application.
+     *
+     * Mapping takes precedence over the default policy resolution.
+     * https://laravel.com/docs/9.x/authorization#policy-auto-discovery
+     *
+     * @var array<class-string, class-string>
+     */
+    protected $policies = [
+        // 'App\Models\Model' => 'App\Policies\ModelPolicy',
+    ];
+
+    /**
      * Register any authentication / authorization services.
+     *
+     * @return void
      */
     public function boot()
     {
         $this->registerPolicies();
-
-        // Policies discovery
-        Gate::guessPolicyNamesUsing(function ($modelClass) {
-            return 'App\Policies\\' . class_basename($modelClass) . 'Policy';
-        });
-
-        Passport::tokensExpireIn((Carbon::now()->addDays(1)));
-        Passport::refreshTokensExpireIn((Carbon::now()->addDays(30)));
     }
 }
