@@ -1,37 +1,21 @@
 <?php
 
+use Illuminate\Support\Facades\Facade;
+
 return [
     /*
     |--------------------------------------------------------------------------
-    | Custom ENVs
+    | (Custom) Application Domain
     |--------------------------------------------------------------------------
     |
-    | Envvars custom created
+    | The application domain is used by the route service provider to assign
+    | a route group to a subdomain. You should set this to the main domain that
+    | your application is running on, without any protocol, application
+    | subdomain or port.
     |
     */
 
     'domain' => env('APP_DOMAIN', 'localhost'),
-    'scheme' => env('APP_SCHEME', 'https'),
-
-    'development_domain' => 'https://carbonaut.io',
-    'production_domain'  => 'https://carbonaut.io',
-
-    /*
-    |--------------------------------------------------------------------------
-    | Config
-    |--------------------------------------------------------------------------
-    |
-    | Custom vars created to use as global constants
-    |
-    */
-
-    'team'               => '',
-    'support_email'      => '',
-    'support_phone'      => '',
-    'mail_header_name'   => '',
-    'mail_header_url'    => '',
-    'mail_header_logo'   => '',
-    'slack_help_channel' => '',
 
     /*
     |--------------------------------------------------------------------------
@@ -57,7 +41,7 @@ return [
     |
     */
 
-    'env' => env('APP_ENV', 'local'),
+    'env' => env('APP_ENV', 'production'),
 
     /*
     |--------------------------------------------------------------------------
@@ -85,7 +69,7 @@ return [
 
     'url' => env('APP_URL', 'http://localhost'),
 
-    'asset_url' => env('ASSET_URL', null),
+    'asset_url' => env('ASSET_URL'),
 
     /*
     |--------------------------------------------------------------------------
@@ -98,8 +82,7 @@ return [
     |
     */
 
-    'timezone'         => 'UTC',
-    'default_timezone' => 'America/Sao_Paulo',
+    'timezone' => 'UTC',
 
     /*
     |--------------------------------------------------------------------------
@@ -112,8 +95,7 @@ return [
     |
     */
 
-    'locale'         => 'de-at',
-    'default_locale' => 'de-at',
+    'locale' => 'en',
 
     /*
     |--------------------------------------------------------------------------
@@ -126,7 +108,7 @@ return [
     |
     */
 
-    'fallback_locale' => 'de-at',
+    'fallback_locale' => 'en',
 
     /*
     |--------------------------------------------------------------------------
@@ -139,7 +121,7 @@ return [
     |
     */
 
-    'faker_locale' => 'en_GB',
+    'faker_locale' => 'en_US',
 
     /*
     |--------------------------------------------------------------------------
@@ -155,6 +137,24 @@ return [
     'key' => env('APP_KEY'),
 
     'cipher' => 'AES-256-CBC',
+
+    /*
+    |--------------------------------------------------------------------------
+    | Maintenance Mode Driver
+    |--------------------------------------------------------------------------
+    |
+    | These configuration options determine the driver used to determine and
+    | manage Laravel's "maintenance mode" status. The "cache" driver will
+    | allow maintenance mode to be controlled across multiple machines.
+    |
+    | Supported drivers: "file", "cache"
+    |
+    */
+
+    'maintenance' => [
+        'driver' => 'file',
+        // 'store'  => 'redis',
+    ],
 
     /*
     |--------------------------------------------------------------------------
@@ -188,7 +188,7 @@ return [
         Illuminate\Redis\RedisServiceProvider::class,
         Illuminate\Auth\Passwords\PasswordResetServiceProvider::class,
         Illuminate\Session\SessionServiceProvider::class,
-        Illuminate\Translation\TranslationServiceProvider::class,
+        Spatie\TranslationLoader\TranslationServiceProvider::class,
         Illuminate\Validation\ValidationServiceProvider::class,
         Illuminate\View\ViewServiceProvider::class,
 
@@ -213,45 +213,7 @@ return [
     |
     */
 
-    'aliases' => [
-        'App'          => Illuminate\Support\Facades\App::class,
-        'Arr'          => Illuminate\Support\Arr::class,
-        'Artisan'      => Illuminate\Support\Facades\Artisan::class,
-        'Auth'         => Illuminate\Support\Facades\Auth::class,
-        'Blade'        => Illuminate\Support\Facades\Blade::class,
-        'Broadcast'    => Illuminate\Support\Facades\Broadcast::class,
-        'Bus'          => Illuminate\Support\Facades\Bus::class,
-        'Cache'        => Illuminate\Support\Facades\Cache::class,
-        'Config'       => Illuminate\Support\Facades\Config::class,
-        'Cookie'       => Illuminate\Support\Facades\Cookie::class,
-        'Crypt'        => Illuminate\Support\Facades\Crypt::class,
-        'Date'         => Illuminate\Support\Facades\Date::class,
-        'DB'           => Illuminate\Support\Facades\DB::class,
-        'Eloquent'     => Illuminate\Database\Eloquent\Model::class,
-        'Event'        => Illuminate\Support\Facades\Event::class,
-        'File'         => Illuminate\Support\Facades\File::class,
-        'Gate'         => Illuminate\Support\Facades\Gate::class,
-        'Hash'         => Illuminate\Support\Facades\Hash::class,
-        'Http'         => Illuminate\Support\Facades\Http::class,
-        'Js'           => Illuminate\Support\Js::class,
-        'Lang'         => Illuminate\Support\Facades\Lang::class,
-        'Log'          => Illuminate\Support\Facades\Log::class,
-        'Mail'         => Illuminate\Support\Facades\Mail::class,
-        'Notification' => Illuminate\Support\Facades\Notification::class,
-        'Password'     => Illuminate\Support\Facades\Password::class,
-        'Queue'        => Illuminate\Support\Facades\Queue::class,
-        'RateLimiter'  => Illuminate\Support\Facades\RateLimiter::class,
-        'Redirect'     => Illuminate\Support\Facades\Redirect::class,
-        'Redis'        => Illuminate\Support\Facades\Redis::class,
-        'Request'      => Illuminate\Support\Facades\Request::class,
-        'Response'     => Illuminate\Support\Facades\Response::class,
-        'Route'        => Illuminate\Support\Facades\Route::class,
-        'Schema'       => Illuminate\Support\Facades\Schema::class,
-        'Session'      => Illuminate\Support\Facades\Session::class,
-        'Storage'      => Illuminate\Support\Facades\Storage::class,
-        'Str'          => Illuminate\Support\Str::class,
-        'URL'          => Illuminate\Support\Facades\URL::class,
-        'Validator'    => Illuminate\Support\Facades\Validator::class,
-        'View'         => Illuminate\Support\Facades\View::class,
-    ],
+    'aliases' => Facade::defaultAliases()->merge([
+        // 'ExampleClass' => App\Example\ExampleClass::class,
+    ])->toArray(),
 ];
