@@ -37,7 +37,7 @@ return [
             'table'        => 'jobs',
             'queue'        => 'default',
             'retry_after'  => 90,
-            'after_commit' => false,
+            'after_commit' => true,
         ],
 
         'beanstalkd' => [
@@ -46,7 +46,7 @@ return [
             'queue'        => 'default',
             'retry_after'  => 90,
             'block_for'    => 0,
-            'after_commit' => false,
+            'after_commit' => true,
         ],
 
         'sqs' => [
@@ -57,7 +57,7 @@ return [
             'queue'        => env('SQS_QUEUE', 'default'),
             'suffix'       => env('SQS_SUFFIX'),
             'region'       => env('AWS_DEFAULT_REGION', 'us-east-1'),
-            'after_commit' => false,
+            'after_commit' => true,
         ],
 
         'redis' => [
@@ -66,7 +66,7 @@ return [
             'queue'        => env('REDIS_QUEUE', 'default'),
             'retry_after'  => 90,
             'block_for'    => null,
-            'after_commit' => false,
+            'after_commit' => true,
         ],
     ],
 
