@@ -1,29 +1,29 @@
 <?php
 
+use App\Enums\Language;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 
-class CreateUsersTable extends Migration
-{
+return new class() extends Migration {
     /**
      * Run the migrations.
+     *
+     * @return void
      */
-    public function up()
+    public function up(): void
     {
         DB::statement('CREATE EXTENSION IF NOT EXISTS "uuid-ossp";');
 
         Schema::create('users', function (Blueprint $table) {
             $table->uuid('id')->primary()->default(DB::raw('uuid_generate_v4()'));
-            $table->string('first_name');
-            $table->string('last_name');
+            $table->string('name');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
-            $table->integer('gender');
-            $table->date('date_of_birth')->nullable();
+            $table->string('language')->default(Language::English->value);
             $table->integer('email_verification_code')->nullable();
             $table->datetime('email_verification_code_expires_at')->nullable();
             $table->timestamps();
@@ -32,9 +32,11 @@ class CreateUsersTable extends Migration
 
     /**
      * Reverse the migrations.
+     *
+     * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('users');
     }
-}
+};
