@@ -9,35 +9,42 @@ class CreateDevicesTable extends Migration
 {
     /**
      * Run the migrations.
+     *
+     * @return void
      */
-    public function up()
+    public function up(): void
     {
         DB::statement('CREATE EXTENSION IF NOT EXISTS "uuid-ossp";');
 
         Schema::create('devices', function (Blueprint $table) {
             $table->uuid('id')->primary()->default(DB::raw('uuid_generate_v4()'));
             $table->uuid('user_id')->index();
+            $table->string('uuid');
+            $table->string('name')->nullable();
             $table->string('platform');
-            $table->string('version');
+            $table->string('operating_system');
+            $table->string('os_version');
             $table->string('manufacturer');
             $table->string('model');
-            $table->string('uuid');
-            $table->string('serial')->nullable();
+            $table->string('web_view_version')->nullable();
             $table->string('app_version')->nullable();
             $table->boolean('is_virtual');
-            $table->boolean('is_active');
             $table->string('push_token')->nullable();
+            $table->boolean('is_active');
             $table->timestamps();
 
-            $table->unique(['user_id', 'uuid', 'platform']);
+            $table->unique(['user_id', 'uuid']);
+
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
     }
 
     /**
      * Reverse the migrations.
+     *
+     * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('devices');
     }
