@@ -8,11 +8,14 @@ class DatabaseSeeder extends Seeder
 {
     /**
      * Seed the application's database.
+     *
+     * @return void
      */
     public function run()
     {
         $this->call([
-            LanguageSeeder::class,
+            UserSeeder::class,
+            LanguageLineSeeder::class,
         ]);
     }
 }
